refactor helper funs into new file & extend getNumOfLines for scalability
