+++ wcTool/.gitignore
@@ -1,0 +1,3 @@
+./.idea
+Ccwc.jar
+./out
